Right now, added de posibility to transfer a mac or name from an user service to an service.
This gives the posibility to to an user service to "reserve" (for now, indefinitely, but will add an "stamp" field to uniqueid database so we can clean up them from time to time) a mac or an name, so other deployed services do not uses them anymore.

One example of this is when a VM gets in error, and machine gets renamed. The mac is them reserved so no more machines clashes with this one reserved.